Add and test opne/close timeslot

+++ controllers/appointment/appointment.ts
@@ -1,6 +1,12 @@
 import { PrismaClient } from "@prisma/client";
 import { Request, Response } from "express";
 import { handleTokenVerification } from "../authentication/authentication";
+import { weekDays } from "../../utils/helps";
+import moment from "moment";
+
+import sgMail from "@sendgrid/mail";
+
+sgMail.setApiKey(process.env.SENDGRID as string);
 
 export const addAppointment = async (
   req: Request,
@@ -13,6 +19,8 @@ export const addAppointment = async (
     const payload: JWTPayload = handleTokenVerification(req, res) as JWTPayload;
 
     if (payload.userId != user_id) return res.status(401).json("Unauthorized");
+
+    if (!moment(date).isValid()) return res.status(400).json("Invalid date");
 
     const doesUserExist = await prisma.user.findUnique({
       where: {
@@ -44,6 +52,49 @@ export const addAppointment = async (
     if (!doesVetExist.is_approved)
       return res.status(400).json("Vet is not approved yet");
 
+    const dateValue = new Date(date);
+    const day = dateValue.getDay() as number;
+
+    const weekDay = (weekDays as any)[day];
+
+    const vetCalender = await prisma.calendar.findUnique({
+      where: {
+        owner_id: vet_id,
+      },
+    });
+
+    if (!JSON.parse((vetCalender as any)[weekDay]))
+      return res.status(400).json("Vet is not available on this day");
+
+    if (parseInt(time.split(":")[1]) % 20 != 0) {
+      return res.status(400).json("Time must be in increments of 20 minutes");
+    }
+    const weekDayWorkingHours = JSON.parse((vetCalender as any)[weekDay]);
+
+    const isBetweenWorkingHours =
+      moment(`${date}  ${time}`).isBetween(
+        `${date}  ${weekDayWorkingHours.morning.start_at}`,
+        `${date}  ${weekDayWorkingHours.morning.end_at}`
+      ) ||
+      moment(`${date}  ${time}`).isBetween(
+        `${date}  ${weekDayWorkingHours.afternoon.start_at}`,
+        `${date}  ${weekDayWorkingHours.afternoon.end_at}`
+      );
+
+    if (!isBetweenWorkingHours)
+      return res.status(400).json("Vet is not available at this time");
+
+    const existingAppointment = await prisma.appointment.findFirst({
+      where: {
+        vet_id,
+        date,
+        time,
+      },
+    });
+
+    if (existingAppointment)
+      return res.status(400).json("Appointment at that time already exists");
+
     const newAppointment = await prisma.appointment.create({
       data: {
         date,
@@ -53,6 +104,23 @@ export const addAppointment = async (
         user_id,
       },
     });
+
+    const msg = {
+      to: doesUserExist.email, // Change to your recipient
+      from: "info@aimensahnoun.com", // Change to your verified sender
+      subject: "VetoLib Appointment",
+      html: `<div><strong>Dear ${doesUserExist.first_name} ${doesUserExist.last_name}</strong> <span> Has booked an appointment with Doctor ${doesVetExist.first_name} ${doesVetExist.last_name} on ${date} ${time} for pet: ${doesPetExist.name}</span></div>
+    `,
+    };
+
+    sgMail
+      .send(msg)
+      .then(() => {
+        console.log("Email sent");
+      })
+      .catch((error) => {
+        console.error(error);
+      });
 
     res.status(200).json(newAppointment);
   } catch (e) {
@@ -67,12 +135,13 @@ export const updateAppointment = async (
 ) => {
   try {
     const { id } = req.params;
-
     const { date, time, pet_id, vet_id, user_id } = req.body;
 
     const payload: JWTPayload = handleTokenVerification(req, res) as JWTPayload;
 
     if (payload.userId != user_id) return res.status(401).json("Unauthorized");
+
+    if (!moment(date).isValid()) return res.status(400).json("Invalid date");
 
     const doesUserExist = await prisma.user.findUnique({
       where: {
@@ -104,7 +173,50 @@ export const updateAppointment = async (
     if (!doesVetExist.is_approved)
       return res.status(400).json("Vet is not approved yet");
 
-    const updatedAppointment = await prisma.appointment.update({
+    const oldAppointment = await prisma.appointment.findUnique({
+      where: {
+        id,
+      },
+    });
+
+    if (!oldAppointment) return res.status(404).json("Appointment not found");
+
+    if (!oldAppointment.user_id == user_id)
+      return res.status(401).json("Unauthorized");
+
+    const dateValue = new Date(date);
+    const day = dateValue.getDay() as number;
+
+    const weekDay = (weekDays as any)[day];
+
+    const vetCalender = await prisma.calendar.findUnique({
+      where: {
+        owner_id: vet_id,
+      },
+    });
+
+    if (!JSON.parse((vetCalender as any)[weekDay]))
+      return res.status(400).json("Vet is not available on this day");
+
+    if (parseInt(time.split(":")[1]) % 20 != 0) {
+      return res.status(400).json("Time must be in increments of 20 minutes");
+    }
+    const weekDayWorkingHours = JSON.parse((vetCalender as any)[weekDay]);
+
+    const isBetweenWorkingHours =
+      moment(`${date}  ${time}`).isBetween(
+        `${date}  ${weekDayWorkingHours.morning.start_at}`,
+        `${date}  ${weekDayWorkingHours.morning.end_at}`
+      ) ||
+      moment(`${date}  ${time}`).isBetween(
+        `${date}  ${weekDayWorkingHours.afternoon.start_at}`,
+        `${date}  ${weekDayWorkingHours.afternoon.end_at}`
+      );
+
+    if (!isBetweenWorkingHours)
+      return res.status(400).json("Vet is not available at this time");
+
+    const newAppointment = await prisma.appointment.update({
       where: {
         id,
       },
@@ -117,7 +229,24 @@ export const updateAppointment = async (
       },
     });
 
-    res.status(200).json(updatedAppointment);
+    const msg = {
+      to: doesUserExist.email, // Change to your recipient
+      from: "info@aimensahnoun.com", // Change to your verified sender
+      subject: "VetoLib Appointment Update",
+      html: `<div><strong>Dear ${doesUserExist.first_name} ${doesUserExist.last_name}</strong> <span> Has booked an appointment with Doctor ${doesVetExist.first_name} ${doesVetExist.last_name} on ${date} ${time} for pet: ${doesPetExist.name}</span></div>
+    `,
+    };
+
+    sgMail
+      .send(msg)
+      .then(() => {
+        console.log("Email sent");
+      })
+      .catch((error) => {
+        console.error(error);
+      });
+
+    res.status(200).json(newAppointment);
   } catch (e) {
     console.log(e);
   }
@@ -169,6 +298,14 @@ export const cancelAppointments = async (
 
     if (payload.userId != user_id) return res.status(401).json("Unauthorized");
 
+    const doesUserExist = await prisma.user.findUnique({
+      where: {
+        id: user_id,
+      },
+    });
+
+    if (!doesUserExist) return res.status(404).json("User not found");
+
     const appointment = await prisma.appointment.delete({
       where: {
         id,
@@ -177,7 +314,214 @@ export const cancelAppointments = async (
 
     if (!appointment) return res.status(404).json("Appointment not found");
 
+    const msg = {
+      to: doesUserExist.email, // Change to your recipient
+      from: "info@aimensahnoun.com", // Change to your verified sender
+      subject: "VetoLib Appointment Cancelled",
+      html: `<div><strong>Dear ${doesUserExist.first_name} ${doesUserExist.last_name}</strong> <span> Has canceled an appointment with id ${appointment.id} successfully </span></div>
+    `,
+    };
+
     res.status(200).json("Appointment canceled successfully");
+  } catch (e) {
+    console.log(e);
+  }
+};
+
+export const closeTimeSlot = async (
+  req: Request,
+  res: Response,
+  prisma: PrismaClient
+) => {
+  try {
+    const { date, time, vet_id , pet_id , user_id } = req.body;
+
+    const payload: JWTPayload = handleTokenVerification(req, res) as JWTPayload;
+
+    if (payload.userId != vet_id) return res.status(401).json("Unauthorized");
+
+    if (!moment(date).isValid()) return res.status(400).json("Invalid date");
+
+    const doesVetExist = await prisma.vet.findUnique({
+      where: {
+        id: vet_id,
+      },
+    });
+
+    if (!doesVetExist)
+      return res.status(404).json(`Vet with id:${vet_id} does not exist`);
+
+    if (!doesVetExist.is_approved)
+      return res.status(400).json("Vet is not approved yet");
+
+    const dateValue = new Date(date);
+    const day = dateValue.getDay() as number;
+
+    const weekDay = (weekDays as any)[day];
+
+    const vetCalender = await prisma.calendar.findUnique({
+      where: {
+        owner_id: vet_id,
+      },
+    });
+
+    if (!JSON.parse((vetCalender as any)[weekDay]))
+      return res.status(400).json("Vet is not available on this day");
+
+    if (parseInt(time.split(":")[1]) % 20 != 0) {
+      return res.status(400).json("Time must be in increments of 20 minutes");
+    }
+    const weekDayWorkingHours = JSON.parse((vetCalender as any)[weekDay]);
+
+    const isBetweenWorkingHours =
+      moment(`${date}  ${time}`).isBetween(
+        `${date}  ${weekDayWorkingHours.morning.start_at}`,
+        `${date}  ${weekDayWorkingHours.morning.end_at}`
+      ) ||
+      moment(`${date}  ${time}`).isBetween(
+        `${date}  ${weekDayWorkingHours.afternoon.start_at}`,
+        `${date}  ${weekDayWorkingHours.afternoon.end_at}`
+      );
+
+    if (!isBetweenWorkingHours)
+      return res.status(400).json("Vet is not available at this time");
+
+    const existingAppointment = await prisma.appointment.findFirst({
+      where: {
+        vet_id,
+        date,
+        time,
+      },
+    });
+
+    if (existingAppointment) {
+      const deletedAppointment = await prisma.appointment.delete({
+        where: {
+          id: existingAppointment.id,
+        },
+      });
+
+      const user = await prisma.user.findUnique({
+        where: {
+          id: existingAppointment.user_id as string,
+        },
+        select: {
+          email: true,
+          first_name: true,
+          last_name: true,
+        },
+      });
+
+      const msg = {
+        to: user?.email, // Change to your recipient
+        from: "info@aimensahnoun.com", // Change to your verified sender
+        subject: "VetoLib Appointment Cancelled By Vet",
+        html: `<div><strong>Dear ${user?.first_name} ${user?.last_name}</strong> <span> Has been canclled by the vet, you are welcome to book another appointment at another time.</span></div>
+      `,
+      };
+
+      sgMail
+        .send(msg)
+        .then(() => {
+          console.log("Email sent");
+        })
+        .catch((error) => {
+          console.error(error);
+        });
+    }
+
+    const newAppointment = await prisma.appointment.create({
+      data: {
+        date,
+        time,
+        pet_id,
+        vet_id,
+        user_id
+      },
+    });
+
+    res.status(200).json(newAppointment);
+  } catch (e) {
+    console.log(e);
+  }
+};
+
+export const openTimeSlot = async (
+  req: Request,
+  res: Response,
+  prisma: PrismaClient
+) => {
+  try {
+    const { date, time, vet_id } = req.body;
+
+    const payload: JWTPayload = handleTokenVerification(req, res) as JWTPayload;
+
+    if (payload.userId != vet_id) return res.status(401).json("Unauthorized");
+
+    if (!moment(date).isValid()) return res.status(400).json("Invalid date");
+
+    const doesVetExist = await prisma.vet.findUnique({
+      where: {
+        id: vet_id,
+      },
+    });
+
+    if (!doesVetExist)
+      return res.status(404).json(`Vet with id:${vet_id} does not exist`);
+
+    if (!doesVetExist.is_approved)
+      return res.status(400).json("Vet is not approved yet");
+
+    const dateValue = new Date(date);
+    const day = dateValue.getDay() as number;
+
+    const weekDay = (weekDays as any)[day];
+
+    const vetCalender = await prisma.calendar.findUnique({
+      where: {
+        owner_id: vet_id,
+      },
+    });
+
+    if (!JSON.parse((vetCalender as any)[weekDay]))
+      return res.status(400).json("Vet is not available on this day");
+
+    if (parseInt(time.split(":")[1]) % 20 != 0) {
+      return res.status(400).json("Time must be in increments of 20 minutes");
+    }
+    const weekDayWorkingHours = JSON.parse((vetCalender as any)[weekDay]);
+
+    const isBetweenWorkingHours =
+      moment(`${date}  ${time}`).isBetween(
+        `${date}  ${weekDayWorkingHours.morning.start_at}`,
+        `${date}  ${weekDayWorkingHours.morning.end_at}`
+      ) ||
+      moment(`${date}  ${time}`).isBetween(
+        `${date}  ${weekDayWorkingHours.afternoon.start_at}`,
+        `${date}  ${weekDayWorkingHours.afternoon.end_at}`
+      );
+
+    if (!isBetweenWorkingHours)
+      return res.status(400).json("Vet is not available at this time");
+
+    const existingAppointment = await prisma.appointment.findFirst({
+      where: {
+        vet_id,
+        date,
+        time,
+      },
+    });
+
+    if (!existingAppointment)
+      return res.status(400).json("Timeslot already open");
+
+    const newAppointment = await prisma.appointment.delete({
+      where: {
+        id: existingAppointment.id,
+      },
+    });
+
+    res.status(200).json(newAppointment);
   } catch (e) {
     console.log(e);
   }
